# openapi.yaml
openapi: '3.0.3'
info:
  title: NGO Partners Backend API
  version: '1.0.0'
  description: |
    REST API for two Flutter apps:
    - **ngo_partner_user** — client-facing app (USER role only)
    - **ngo_partner_sales** — internal CRM app (SALES + ADMIN roles)

    ## Identity Model
    There is a single `USER` role for all clients. NGO context is stored as
    profile fields (`isOrganizationAccount`, `organizationName`, etc.) rather
    than a separate role. An NGO is simply a USER whose `isOrganizationAccount = true`.

    ## Authentication
    - JWT Bearer tokens (15 min access / 7 day refresh)
    - Refresh token rotation — single use, stored as SHA-256 hash
    - Brute-force lockout after 5 failed attempts (15 min window)
servers:
  - url: http://localhost:4000/api/v1
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /reports/revenue:
    get:
      tags: [Reports]
      summary: List revenue records (SALES/ADMIN)
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Revenue records
  /reports/export:
    post:
      tags: [Reports]
      summary: Request a report export (SALES/ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reportType]
              properties:
                reportType:
                  type: string
                  enum: [revenue, leads, cases, tickets]
      responses:
        '201':
          description: Export record created

  # ── Audit ──────────────────────────────────────────────────────────────────────
...